refactor: Move downloadedSur feature to quran package
