+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/quran/downloadedSur/DownloadedSurEffect.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.faith.presentation.feature.downloadedSur
+package net.thechance.mena.faith.presentation.feature.quran.downloadedSur
 
 sealed interface DownloadedSurEffect {
     data object NavigateBack : DownloadedSurEffect
