+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/navigation/FaithNavGraph.kt
@@ -8,7 +8,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import net.thechance.mena.faith.presentation.designSystem.theme.QuranTheme
-import net.thechance.mena.faith.presentation.feature.downloadedSur.DownloadedSurScreen
+import net.thechance.mena.faith.presentation.feature.quran.downloadedSur.DownloadedSurScreen
 import net.thechance.mena.faith.presentation.feature.main.MainScreen
 import net.thechance.mena.faith.presentation.feature.mosque.NearbyMosquesScreen
 import net.thechance.mena.faith.presentation.feature.prayertime.PrayerTimeScreen
